Updated to object to use spread operator
Here we're just doing a different method than the previous commit which is a bit less verbose. This is essentially accomplishing the same thing as before in terms of passing through an entire object as a prop, however, we're utilizing the spread operator which just changes how we call each of the props in our Card.js component.